# src/main/java/com/bank/anomaly/controller/TransactionController.java
package com.bank.anomaly.controller;

import com.bank.anomaly.model.EvaluationResult;
import com.bank.anomaly.model.Transaction;
import com.bank.anomaly.repository.RiskResultRepository;
import com.bank.anomaly.service.TransactionEvaluationService;
import com.bank.anomaly.service.TransactionService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/v1/transactions")
@Tag(name = "Transactions", description = "Submit transactions for anomaly evaluation and query transaction history")
public class TransactionController {

    private final TransactionEvaluationService evaluationService;
    private final RiskResultRepository riskResultRepository;
    private final TransactionService transactionService;

    public TransactionController(TransactionEvaluationService evaluationService,
                                 RiskResultRepository riskResultRepository,
                                 TransactionService transactionService) {
        this.evaluationService = evaluationService;
        this.riskResultRepository = riskResultRepository;
        this.transactionService = transactionService;
    }

    @Operation(summary = "Evaluate a transaction for anomalies",
            description = "Submits a transaction for real-time anomaly evaluation against all active rules " +
                    "(including the Isolation Forest ML model). Returns composite risk score, risk level, " +
                    "action (PASS/ALERT/BLOCK), and per-rule breakdown.")
    @PostMapping("/evaluate")
    public ResponseEntity<EvaluationResult> evaluateTransaction(@RequestBody Transaction txn) {
        if (txn.getTxnId() == null || txn.getClientId() == null || txn.getTxnType() == null) {
            return ResponseEntity.badRequest().build();
        }

        if (txn.getTimestamp() == 0) {
            txn.setTimestamp(System.currentTimeMillis());
        }

        EvaluationResult result = evaluationService.evaluate(txn);
        return ResponseEntity.ok(result);
    }

    @Operation(summary = "Get a transaction by ID",
            description = "Retrieves a persisted transaction by its unique transaction ID.")
    @GetMapping("/{txnId}")
    public ResponseEntity<Transaction> getTransaction(
            @Parameter(description = "Transaction ID", example = "CLIENT-001-TXN-000001")
            @PathVariable String txnId) {
        Transaction txn = transactionService.getTransaction(txnId);
        if (txn == null) {
            return ResponseEntity.notFound().build();
        }
        return ResponseEntity.ok(txn);
    }

    @Operation(summary = "List transactions by client ID",
            description = "Retrieves recent transactions for a specific client, ordered by timestamp.")
    @GetMapping("/client/{clientId}")
    public ResponseEntity<List<Transaction>> getTransactionsByClient(
            @Parameter(description = "Client ID", example = "CLIENT-001")
            @PathVariable String clientId,
            @Parameter(description = "Max number of transactions to return", example = "50")
            @RequestParam(defaultValue = "50") int limit) {
        List<Transaction> txns = transactionService.getTransactionsByClientId(clientId, limit);
        return ResponseEntity.ok(txns);
    }

    @Operation(summary = "Get evaluation result for a transaction",
            description = "Retrieves the anomaly evaluation result for a previously evaluated transaction.")
    @GetMapping("/results/{txnId}")
    public ResponseEntity<EvaluationResult> getResult(
            @Parameter(description = "Transaction ID", example = "IF-DEMO-001")
            @PathVariable String txnId) {
        EvaluationResult result = riskResultRepository.findByTxnId(txnId);
        if (result == null) {
            return ResponseEntity.notFound().build();
        }
        return ResponseEntity.ok(result);
    }

    @Operation(summary = "List evaluation results by client ID",
            description = "Retrieves recent evaluation results for a specific client.")
    @GetMapping("/results/client/{clientId}")
    public ResponseEntity<List<EvaluationResult>> getResultsByClient(
            @Parameter(description = "Client ID", example = "CLIENT-001")
            @PathVariable String clientId,
            @Parameter(description = "Max number of results to return", example = "20")
            @RequestParam(defaultValue = "20") int limit) {
        List<EvaluationResult> results = riskResultRepository.findByClientId(clientId, limit);
        return ResponseEntity.ok(results);
    }
}